revert: commit a95985f partially.
The commit a95985f removed the declaration of the `ThemeServiceProvider`. This is not needed for an upcoming version of Sage, but is needed for Sage 10.8.2.

Fixes PHP/Symfony error: "Target class [sage.view] does not exist." by re-adding the declaration of `ThemeServiceProvider` in composer.json.

Related to commit 8a9c7c2.